+++ examples/wallet/src/CreateIdentity.tsx
@@ -1,54 +1,10 @@
-import React, { useEffect, useMemo, useState } from 'react';
-import { identityIndex, network, seedPhraseCookie } from './Index';
+import React, { useEffect, useState } from 'react';
+import { network, seedPhraseCookie } from './Index';
 import { useNavigate } from 'react-router-dom';
-import { CryptographicParameters, IdObjectRequestV1, IdentityRequestInput, Network, Versioned, createIdentityRequest } from '@concordium/web-sdk';
-import { mnemonicToSeedSync } from '@scure/bip39';
-import { Buffer } from 'buffer/';
+import { CryptographicParameters } from '@concordium/web-sdk';
 import { IdentityProviderWithMetadata } from './types';
 import { useCookies } from 'react-cookie';
-import { getCryptographicParameters, getIdentityProviders } from './util';
-
-const redirectUri = 'http://localhost:4173/identity';
-
-function createIdentityObjectRequest(identityProvider: IdentityProviderWithMetadata, global: CryptographicParameters, network: Network, seedPhrase: string) {
-    const seedCorrectFormat = Buffer.from(mnemonicToSeedSync(seedPhrase)).toString('hex');
-
-    const identityRequestInput: IdentityRequestInput = {
-        net: network,
-        seed: seedCorrectFormat,
-        identityIndex: identityIndex,
-        arsInfos: identityProvider.arsInfos,
-        // TODO Explain why this is chosen like this.
-        arThreshold: Math.min(Object.keys(identityProvider.arsInfos).length - 1, 255),
-        ipInfo: identityProvider.ipInfo,
-        globalContext: global
-    };
-    return createIdentityRequest(identityRequestInput);
-}
-
-function buildURLwithSearchParameters(baseUrl: string, params: Record<string, string>) {
-    const searchParams = new URLSearchParams(params);
-    return Object.entries(params).length === 0 ? baseUrl : `${baseUrl}?${searchParams.toString()}`;
-}
-
-async function sendRequest(idObjectRequest: Versioned<IdObjectRequestV1>, baseUrl: string) {
-    const params = {
-        scope: 'identity',
-        response_type: 'code',
-        redirect_uri: redirectUri,
-        state: JSON.stringify({ idObjectRequest }),
-    };
-
-    const url = buildURLwithSearchParameters(baseUrl, params);
-    const response = await fetch(url);
-
-    // The identity creation protocol dictates that we will receive a redirect.
-    if (!response.redirected) {
-        // Something went wrong...
-    } else {
-        return response.url;
-    }
-}
+import { createIdentityObjectRequest, getCryptographicParameters, getIdentityProviders, redirectUri, sendRequest } from './util';
 
 export function CreateIdentity() {
     const [identityProviders, setIdentityProviders] = useState<IdentityProviderWithMetadata[]>();
@@ -83,9 +39,10 @@ export function CreateIdentity() {
             return;
         }
 
-        const request = createIdentityObjectRequest(selectedIdentityProvider, cryptographicParameters, network, seedPhrase);
-        const url = await sendRequest(request, selectedIdentityProvider.metadata.issuanceStart);
+        const identityObjectRequest = createIdentityObjectRequest(selectedIdentityProvider, cryptographicParameters, network, seedPhrase);
+        const url = await sendRequest(identityObjectRequest, selectedIdentityProvider.metadata.issuanceStart);
 
+        // TODO Explain why this check is required.
         if (!url?.includes(redirectUri)) {
             window.open(url);
         }
